Move and consolidate some files
App has been fused with index.js, to create a cleaner file structure.
I'm sure components shouldn't be mixed with root logic, but oh well.

Add Typist for typing some text
