+++ .eslintrc
@@ -3,7 +3,8 @@
   "rules": {
     "react/jsx-filename-extension": 0,
     "import/no-extraneous-dependencies": ["error", { "devDependencies": true }],
-    "import/no-named-as-default": 0
+    "import/no-named-as-default": 0,
+    "linebreak-style": 0
   },
   "globals": {
     "document": true,
