+++ package.json
@@ -17,6 +17,7 @@
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
     "react-scripts": "1.0.17",
+    "react-typist": "^2.0.4",
     "shortid": "^2.2.8",
     "styled-components": "^2.4.0"
   },
