+++ src/sections/home.css
@@ -26,7 +26,7 @@
   margin-top: -10px;
 }
 
-.slogan p {
+.slogan * {
   text-align: center;
   margin: 0;
   font-size: 1.5rem;
