+++ src/sections/home.js
@@ -1,4 +1,5 @@
 import React from "react";
+import Typist from "react-typist";
 
 import imgSpeechBubble from "../images/home-speechbubble.png";
 import imgNamePlate from "../images/home-nameplate.svg";
@@ -12,8 +13,20 @@ export default () => (
       <img className="logo" src={imgNamePlate} alt="Heather Allard" />
     </div>
     <div className="slogan">
-      <p>
+      <Typist
+        className="typing"
+        startDelay={2000}
+        cursor={{ hideWhenDone: true }}
+      >
+        eat.
+        <Typist.Backspace count={4} delay={1250} />
+        design.
+        <Typist.Backspace count={7} delay={1250} />
+        adventure.
+        <Typist.Backspace count={11} delay={1000} />
         eat. design. adventure. repeat.
+      </Typist>
+      <p>
         <br />what can I do for you?
       </p>
     </div>
